+++ installer.cfg
@@ -1,7 +1,7 @@
 [Application]
 name = UCI CBP DEMO
 version = 0.4.0
-entry_point = uci_cbp_demo:gui_noargs
+entry_point = uci_cbp_demo:main
 icon = pulse_8C8_icon.ico
 publisher = Michael Tao-Yi Lee
