+++ uci_cbp_demo/backend/bluetooth/__init__.py
@@ -4,4 +4,3 @@
 from .callbacks import CapData
 from .list_characteristics import list_char
 from .scan import scan, check_adaptor
-from .sensorBoard import SensorBoard
