+++ uci_cbp_demo/backend/bluetooth/constants.py
