+++ uci_cbp_demo/backend/bluetooth/list_characteristics.py
