+++ uci_cbp_demo/logging/__init__.py
@@ -1,5 +1,4 @@
 #  MIT License
 #  Copyright (C) Michael Tao-Yi Lee (taoyil AT UCI EDU)
 
-from .gui import main
-from .tui import tui_main
+from .logger import logger, console_handler
