+++ uci_cbp_demo/logging/logger.py
@@ -1,0 +1,11 @@
+#  MIT License
+#  Copyright (C) Michael Tao-Yi Lee (taoyil AT UCI EDU)
+import logging
+
+logger = logging.getLogger("bp_demo")
+console_handler = logging.StreamHandler()
+logger.setLevel(logging.INFO)
+console_handler.setLevel(logging.INFO)
+formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
+console_handler.setFormatter(formatter)
+logger.addHandler(console_handler)
